# AI_Math_Project/AI_Math_Project/AI_MATH_PROJECT/AI_Math_Project/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Math AI Api",
    "description": "This is all api services provide for AI Math product",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Email",
      "email": "trongleviet05@gmail.com"
    },
    "license": {
      "name": "License",
      "url": "https://example.com/license"
    },
    "version": "v1"
  },
  "paths": {
    "/api/chapters": {
      "get": {
        "tags": [
          "Chapter"
        ],
        "summary": "Return all chapters of all class",
        "description": "- **grade**: The grade level\r\n- **chapterOrder**: The order of the chapter in the curriculum.\r\n- **chapterName**: The name of the chapter.\r\n- **lessons**: Null, because this api just return information about grade and chapters, not include lessions in chapter",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ChapterDto"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/chapters/details": {
      "get": {
        "tags": [
          "Chapter"
        ],
        "summary": "Returns detailed information of each lesson in the chapters",
        "description": "- **grade**: The grade level\r\n- **chapterOrder**: The order of the chapter in the curriculum.\r\n- **semester**: semester\r\n- **chapterName**: The name of the chapter.\r\n- **lessons**: A list of lessons within the chapter. Each lesson includes:\r\n  - **lessonOrder**: The order of the lesson within the chapter.\r\n  - **lessonName**: The name of the lesson.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ChapterDto"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/chapters/grade/{grade}/details": {
      "get": {
        "tags": [
          "Chapter"
        ],
        "summary": "Returns detailed information of each lesson in the chapters classified by class",
        "description": "- **grade**: The grade level\r\n- **chapterOrder**: The order of the chapter in the curriculum.\r\n- **chapterName**: The name of the chapter.\r\n- **semester**: semester \r\n- **lessons**: A list of lessons within the chapter. Each lesson includes:\r\n  - **lessonOrder**: The order of the lesson within the chapter.\r\n  - **lessonName**: The name of the lesson.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "grade",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ChapterDto"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/enrollment/id/{id}": {
      "get": {
        "tags": [
          "Enrollment"
        ],
        "summary": "Retrieves all enrollment information for a specific user.",
        "description": "This API retrieves enrollment records for a given user, including details such as grade, semester, and academic year.\r\n            \r\n**Request:**\r\nThe request should include:\r\n- **id** (int): The unique identifier of the user.\r\n            \r\n**Response:**\r\nThe response will return a list of enrollment records, each containing:\r\n- **enrollmentId**: The unique identifier of the enrollment.\r\n- **userId**: The unique identifier of the user.\r\n- **grade**: The grade level associated with the enrollment.\r\n- **enrollmentDate**: The date when the user enrolled.\r\n- **avgScore**: The average score of the user in this enrollment (nullable).\r\n- **semester**: The semester in which the user is enrolled.\r\n- **startYear**: The academic start year.\r\n- **endYear**: The academic end year.\r\n            \r\n**Example Request:**\r\n```http\r\nGET /getinfo/id/10",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/EnrollmentDto"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/lesson/grade/{grade}/chapterorder/{chapterorder}/lessonorder/{lessonorder}": {
      "get": {
        "tags": [
          "Lession"
        ],
        "summary": "Retrieves the details of a lesson associated with a specific grade, chapter, and lesson order.",
        "description": "This API returns the details of a lesson, filtered by grade level, chapter order, and lesson order.\r\n\r\n**Request Parameters:**\r\n- **grade** (int): The grade level of the study program.\r\n- **chapterorder** (int): The sequential order of the chapter.\r\n- **lessonorder** (int): The sequential order of the lesson within the chapter.\r\n\r\n**Response Format:**\r\nThe response will return the details of the lesson, including:\r\n- **lessonOrder** (short?): The order of the lesson within the chapter. Nullable to allow for optional lesson orders.\r\n- **lessonName** (string): The name of the lesson.\r\n- **lessonContent** (string, nullable): The content of the lesson, if available. This can be null if no content is provided.\r\n\r\n**Example Request:**\r\n```http\r\nGET /api/lesson/grade/1/chapterorder/1/lessonorder/1\r\n```",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "grade",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "chapterorder",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "lessonorder",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LessonDto"
            }
          }
        }
      }
    },
    "/api/lesson/grade/{grade}/chapter/{chapterorder}": {
      "post": {
        "tags": [
          "Lession"
        ],
        "summary": "Creates a new lesson associated with a specific grade and chapter order.",
        "description": "This API creates a new lesson and saves it in the system. The lesson is created with the provided grade level, chapter order, and lesson data.\r\n\r\n**Request Parameters:**\r\n- **grade** (int): The grade level of the study program.\r\n- **chapterorder** (int): The sequential order of the chapter in the study program.\r\n- **lessonDto** (LessonDto): The details of the lesson to be created. This should include:\r\n  - **lessonOrder** (short?): The order of the lesson within the chapter.\r\n  - **lessonName** (string): The name of the lesson.\r\n  - **lessonContent** (string, nullable): The content of the lesson (optional).\r\n\r\n**Response Format:**\r\n- **201 Created**: If the lesson is successfully created, the response will include the created `LessonDto` and the URL of the new lesson.\r\n- **400 Bad Request**: If the lesson could not be created, the response will contain an error message with details on the failure.\r\n\r\n**Example Request:**\r\n```http\r\nPOST /api/lesson/grade/1/chapter/1\r\nContent-Type: application/json\r\n{\r\n  \"lessonOrder\": 2,\r\n  \"lessonName\": \"Vị trí tiết 2\",\r\n  \"lessonContent\": \"linkyoutube\"\r\n}\r\n**Example Response (Success):**\r\n```json\r\nHTTP/1.1 201 Created\r\nLocation: /api/lesson/grade/1/chapter/1/lessonorder/2\r\n{\r\n  \"lessonOrder\": 2,\r\n  \"lessonName\": \"Vị trí tiết 2\",\r\n  \"lessonContent\": \"linkyoutube\"\r\n}",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "grade",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "chapterorder",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LessonDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LessonDto"
            }
          }
        }
      }
    },
    "/id/{id}": {
      "get": {
        "tags": [
          "LessionProgress"
        ],
        "summary": "Returns all information of the study program the user has registered for.",
        "description": "This API retrieves the user's study program information, including lesson details and learning progress.\r\n\r\n**Request:**\r\nThe request should include userID.\r\n\r\n**Response:**\r\nThe response will return a list of lessons, each containing:\r\n- **lessonId**: The unique identifier of the lesson.\r\n- **learningProgress**: The progress of the user in the lesson (e.g., percentage completed).\r\n- **isCompleted**: A boolean indicating whether the lesson is completed.\r\n- **lesson**: An object containing lesson details:\r\n  - **lessonOrder**: The order of the lesson in the study plan.\r\n  - **lessonName**: The name of the lesson.\r\n  - **lessonContent**: The content of the lesson.\r\n **Example Request:**\r\n```http\r\nGET /api/lessionprogress/id/6",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/LessonProgressDto"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/id/{id}/semester/{semester}": {
      "get": {
        "tags": [
          "LessionProgress"
        ],
        "summary": "Returns all information of the study program classified by the semester the user has registered for.",
        "description": "This API retrieves the user's study program information, including lesson details and learning progress.\r\n\r\n**Request:**\r\nThe request should include **userID** and **Semester**.\r\n\r\n**Response:**\r\nThe response will return a list of lessons, each containing:\r\n- **lessonId**: The unique identifier of the lesson.\r\n- **learningProgress**: The progress of the user in the lesson (e.g., percentage completed).\r\n- **isCompleted**: A boolean indicating whether the lesson is completed.\r\n- **lesson**: An object containing lesson details:\r\n  - **lessonOrder**: The order of the lesson in the study plan.\r\n  - **lessonName**: The name of the lesson.\r\n  - **lessonContent**: The content of the lesson.\r\n  **Example Request:**\r\n```http\r\nGET /api/lessionprogress/id/6/semester/1",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "semester",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/LessonProgressDto"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/update/lessionprogressID/{idProgress}/learningprogress/{learningProgress}": {
      "patch": {
        "tags": [
          "LessionProgress"
        ],
        "summary": "Updates the learning progress of a specific lesson progress record.",
        "description": "This API updates the user's learning progress for a specific lesson.\r\n\r\n**Request:**\r\nThe request should include:\r\n- **idProgress** (int): The unique identifier of the lesson progress.\r\n- **learningProgress** (short): The updated progress percentage of the lesson.\r\n            \r\n**Response:**\r\nIf successful, the response will return the updated lesson progress information:\r\n- **learningProgressId**: The unique identifier of the lesson progress.\r\n- **lessonId**: The unique identifier of the lesson.\r\n- **learningProgress**: The updated progress of the user in the lesson (e.g., percentage completed).\r\n- **isCompleted**: A boolean indicating whether the lesson is completed.\r\n- **lesson**: An object containing lesson details:\r\n  - **lessonOrder**: The order of the lesson in the study plan.\r\n  - **lessonName**: The name of the lesson.\r\n  - **lessonContent**: A link to the lesson content.\r\n            \r\n**Example Request:**\r\n```http\r\nPATCH /update/lessionprogressID/13/learningprogress/23",
        "parameters": [
          {
            "in": "path",
            "name": "idProgress",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "learningProgress",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/grade/{grade}/lessionorder/{lessionorder}": {
      "get": {
        "tags": [
          "Question"
        ],
        "summary": "Retrieves all questions associated with a specific grade and lesson order.",
        "description": "This API returns a list of questions for a given study program, filtered by grade level and lesson order.\r\n            \r\n**Request Parameters:**\r\n- **grade** (int): The grade level of the study program.\r\n- **lessionOrder** (int): The sequential order of the lesson.\r\n            \r\n**Response Format:**\r\nThe response will return a list of questions, each containing:\r\n- **questionId** (int): The unique identifier of the question.\r\n- **questionType** (string): The type of question (e.g., \"multiple_choice\", \"fill_in_blank\", \"matching\").\r\n- **difficulty** (int): The difficulty level of the question.\r\n- **lessonId** (int): The identifier of the associated lesson.\r\n- **imgUrl** (string, nullable): The URL of an image related to the question.\r\n- **questionContent** (string): The text content of the question.\r\n- **pdfSolution** (string, nullable): A PDF solution reference, if available.\r\n- **choiceAnswers** (array): List of multiple-choice answers (if applicable):\r\n  - **answerId** (int): The unique identifier of the answer choice.\r\n  - **content** (string, nullable): The text of the answer choice (if applicable).\r\n  - **isCorrect** (bool): Whether this choice is the correct answer.\r\n  - **imgUrl** (string, nullable): An optional image URL for the answer choice.\r\n- **fillAnswers** (array): List of correct answers for fill-in-the-blank questions (if applicable):\r\n  - **answerId** (int): The unique identifier of the correct answer.\r\n  - **correctAnswer** (string): The correct text answer that should be filled in.\r\n  - **position** (short): The position of the blank in the question (for multi-blank questions).\r\n- **matchingAnswers** (array): List of matching question pairs (if applicable):\r\n  - **answerId** (int): The unique identifier of the matching answer.\r\n  - **correctAnswer** (string): The correct match for the question.\r\n  - **imgUrl** (string, nullable): An image URL representing the matching item.\r\n            \r\n**Example Request:**\r\n```http\r\nGET /api/question/grade/1/lessionorder/1\r\n```",
        "parameters": [
          {
            "in": "path",
            "name": "grade",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "lessionorder",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "ChapterDto": {
      "type": "object",
      "properties": {
        "grade": {
          "format": "int32",
          "type": "integer"
        },
        "chapterOrder": {
          "format": "int32",
          "type": "integer"
        },
        "chapterName": {
          "type": "string"
        },
        "semester": {
          "format": "int32",
          "type": "integer"
        },
        "lessons": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LessonDto"
          }
        }
      },
      "additionalProperties": false
    },
    "EnrollmentDto": {
      "type": "object",
      "properties": {
        "enrollmentId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "grade": {
          "format": "int32",
          "type": "integer"
        },
        "enrollmentDate": {
          "format": "date",
          "type": "string"
        },
        "avgScore": {
          "format": "double",
          "type": "number"
        },
        "semester": {
          "format": "int32",
          "type": "integer"
        },
        "startYear": {
          "format": "int32",
          "type": "integer"
        },
        "endYear": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "LessonDto": {
      "type": "object",
      "properties": {
        "lessonOrder": {
          "format": "int32",
          "type": "integer"
        },
        "lessonName": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "lessonContent": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "LessonProgressDto": {
      "type": "object",
      "properties": {
        "learningProgressId": {
          "format": "int32",
          "type": "integer"
        },
        "lessonId": {
          "format": "int32",
          "type": "integer"
        },
        "learningProgress": {
          "format": "int32",
          "type": "integer"
        },
        "isCompleted": {
          "type": "boolean"
        },
        "lesson": {
          "$ref": "#/definitions/LessonDto"
        }
      },
      "additionalProperties": false
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    }
  }
}